+++ changelog.md
@@ -5,7 +5,11 @@
 ### Components changes
 
 Updated images for components `mash_dist`, `mash_screen` and 
-`mapping_patlas`
+`mapping_patlas`.
+
+### New components
+- Added component `fast_ani`.
+
 
 ## 1.3.0